Add files via upload
                                                         # LAB ONE
i. WAP to input name, roll number and marks in 5 subjects for a student, and display it.
ii. WAP to input name, roll number and marks in 5 subjects for n number of students. Write
functions to:-
a. Find total marks and percentage of all n students.
b. Display details of a student with a given roll number.
c. Display the details for all the students having percentage in a given range.
d. Sort the array in ascending order of marks.