Add files via upload
i. WAP to input name, roll number and marks in 5 subjects for a student, and display it.
ii. WAP to input name, roll number and marks in 5 subjects for n number of students. Write
functions to:-
a. Find total marks and percentage of all n students.
b. Display details of a student with a given roll number.
c. Display the details for all the students having percentage in a given range.
d. Sort the array in ascending order of marks.
iii. WAP to enter id, name, age and basic salary of n number of employees. Calculate the
gross salary of all the employees and display it along with all other details in a tabular
form, using pointer to structure.
[ Gross salary= Basic salary + DA + HRA,
DA = 80% of Basic salary
HRA=10% of Basic salary ]