Add files via upload
Q1.WAP to perform the addition of two integers and display the result.
Q2 WAP to find centigrade for a given Fahrenheit temperature.
Q3. WAP to calculate area of a circle.
Q4. WAP to calculate area of a triangle whose base and height is given.
Q5 WAP to swap two integer numbers using third variable.
Q6. WAP to find the largest between two numbers.
Q7. WAP to read an alphabet from the user and convert it into uppercase if the entered alphabet is in
lowercase, otherwise display an appropriate message.
Q8. WAP to input any two integers, and provide a menu to the user to select any of the options as add,
subtract, multiply, divide and display the result accordingly.
Q9. WAP to print the natural numbers from 1 to 20.
Q10. WAP to print all numbers within a given range. The range is given by user.
Q11. WAP to calculate the factorial of a given number
Q12. WAP to calculate the sum of digits of a given number
Q13. WAP to check whether a number n is prime number or not.